+++ api/base.py
@@ -1,0 +1,49 @@
+from collections import OrderedDict
+
+from rest_framework_json_api import serializers
+
+from api import fields
+
+
+class ShareSerializer(serializers.ModelSerializer):
+    id = fields.ObfuscatedIDField()
+
+    # http://stackoverflow.com/questions/27015931/remove-null-fields-from-django-rest-framework-response
+    def to_representation(self, instance):
+        ret = super().to_representation(instance)
+        ret = OrderedDict(list(filter(lambda x: x[1] is not None, ret.items())))
+        return ret
+
+
+class ShareObjectSerializer(ShareSerializer):
+
+    def __init__(self, *args, **kwargs):
+        # super hates my additional kwargs
+        sparse = kwargs.pop('sparse', False)
+        version_serializer = kwargs.pop('version_serializer', False)
+        super().__init__(*args, **kwargs)
+
+        if sparse:
+            # clear the fields if they asked for sparse
+            self.fields.clear()
+        else:
+            # remove hidden fields
+            excluded_fields = ['change', 'sources']
+            for field_name in tuple(self.fields.keys()):
+                if 'version' in field_name or field_name in excluded_fields:
+                    self.fields.pop(field_name)
+
+        # version specific fields
+        if version_serializer:
+            self.fields.update({
+                'action': serializers.CharField(max_length=10),
+                'persistent_id': serializers.IntegerField()
+            })
+
+        # add fields with improper names
+        self.fields.update({
+            'type': fields.TypeField(),
+        })
+
+    class Meta:
+        links = ('versions', 'changes', 'rawdata')
