+++ api/fields.py
@@ -1,6 +1,7 @@
-import shortuuid
 from rest_framework import serializers
-from rest_framework.reverse import reverse
+
+from share.util import IDObfuscator
+from share.util import InvalidID
 
 
 class TypeField(serializers.ReadOnlyField):
@@ -14,52 +15,18 @@ class TypeField(serializers.ReadOnlyField):
         return obj._meta.model_name
 
 
-class ObjectIDField(serializers.ReadOnlyField):
-    """
-    Returns a cleaner version of a uuid
-    """
+class ObfuscatedIDField(serializers.ReadOnlyField):
 
     def get_attribute(self, instance):
         return instance
 
-    def to_representation(self, value):
-        shortuuid.set_alphabet('23456789abcdefghjkmnpqrstuvwxyz')
-        return shortuuid.encode(value.uuid)
+    def to_representation(self, instance):
+        return IDObfuscator.encode(instance)
 
+    def to_internal_value(self, value):
+        import ipdb; ipdb.set_trace()
 
-class DetailUrlField(serializers.ReadOnlyField):
-    """
-    Returns an object's @id, the url to its detail page.
-    e.g. 'http://share.osf.io/api/v2/preprints/18'
-    """
-
-    def get_attribute(self, instance):
-        return instance
-
-    def to_representation(self, obj):
-        model = obj._meta.model_name
-        if 'version' in model:
-            model = model[:-7]
-        view = 'api:{}-detail'.format(model)
-        request = self.context['request']
-        return reverse(view, kwargs={'pk': obj.id}, request=request)
-
-
-class LinksField(serializers.ReadOnlyField):
-    """
-    Returns a dictionary of links to an object's relationships.
-    """
-
-    def __init__(self, **kwargs):
-        self._links = kwargs.pop('links', ())
-        super(serializers.ReadOnlyField, self).__init__(**kwargs)
-
-    def get_attribute(self, instance):
-        return instance
-
-    def to_representation(self, obj):
-        request = self.context['request']
-        return {
-            l: reverse('api:{}-{}'.format(obj._meta.model_name, l), kwargs={'pk': obj.id}, request=request)
-            for l in self._links
-        }
+        try:
+            return IDObfuscator.decode_id(value)[1]
+        except InvalidID:
+            raise serializers.ValidationError('Invalid ID')
