+++ api/util.py
@@ -1,0 +1,9 @@
+from django import http
+
+
+class HttpSmartResponseRedirect(http.HttpResponseRedirect):
+    status_code = 307
+
+
+class HttpSmartResponsePermanentRedirect(http.HttpResponsePermanentRedirect):
+    status_code = 308
