+++ requirements.txt
@@ -37,7 +37,6 @@ pycountry==1.20  # LGPL 2.1
 pyyaml==3.11  # MIT
 raven==6.1.0  # BSD
 requests==2.10.0  # Apache 2.0
-shortuuid==0.4.3  # BSD
 stevedore==1.21.0  # Apache 2.0
 ujson==1.35  # BSD
 werkzeug==0.11.10  # BSD
